requirements.txt
openai>=1.75.0
pydantic>=2.11.0
PyMuPDF>=1.25.0
python-dotenv>=1.0.1
